added a new file and edited estimator
